feat(api): Rename accreditation-controller to complementary-certification-controller
