+++ api/lib/application/complementary-certifications/complementary-certification-controller.js
@@ -2,7 +2,7 @@ const usecases = require('../../domain/usecases');
 const accreditationSerializer = require('../../infrastructure/serializers/jsonapi/accreditation-serializer');
 
 module.exports = {
-  async findAccreditations() {
+  async findComplementaryCertifications() {
     const complementaryCertifications = await usecases.findComplementaryCertifications();
     return accreditationSerializer.serialize(complementaryCertifications);
   },
