+++ api/tests/unit/application/complementary-certifications/index_test.js
@@ -1,16 +1,16 @@
 const { expect, HttpTestServer, sinon } = require('../../../test-helper');
-const moduleUnderTest = require('../../../../lib/application/accreditations');
-const accreditationController = require('../../../../lib/application/accreditations/accreditation-controller');
+const moduleUnderTest = require('../../../../lib/application/complementary-certifications');
+const complementaryCertificationController = require('../../../../lib/application/complementary-certifications/complementary-certification-controller');
 const securityPreHandlers = require('../../../../lib/application/security-pre-handlers');
 
-describe('Unit | Application | Router | accreditations-router', function () {
+describe('Unit | Application | Router | complementary-certifications-router', function () {
   describe('GET /api/accreditations', function () {
     it('should return 403 HTTP status code when the user authenticated is not PixMaster', async function () {
       // given
       sinon
         .stub(securityPreHandlers, 'checkUserIsAdminInOrganization')
         .callsFake((request, h) => h.response().code(403).takeover());
-      sinon.stub(accreditationController, 'findAccreditations').returns('ok');
+      sinon.stub(complementaryCertificationController, 'findComplementaryCertifications').returns('ok');
       const httpTestServer = new HttpTestServer();
       await httpTestServer.register(moduleUnderTest);
 
@@ -19,7 +19,7 @@ describe('Unit | Application | Router | accreditations-router', function () {
 
       // then
       expect(response.statusCode).to.equal(403);
-      sinon.assert.notCalled(accreditationController.findAccreditations);
+      sinon.assert.notCalled(complementaryCertificationController.findComplementaryCertifications);
     });
   });
 });
